Moved tests from JwtValidatorFactoryTest to OAuthBearerValidatorCallbackHandlerTest
